Move placeholder content from Main up to Index.
