+++ pages/index.tsx
@@ -13,15 +13,33 @@ import { WorkInProgress } from '../components/WorkInProgress'
 
 import queryApiGateway from '../utils/queryApiGateway'
 import getScenarioContent from '../utils/getScenarioContent'
-import { Claim } from '../types/common'
+import { ClaimDetailsContent, ClaimStatusContent, ScenarioContent } from '../types/common'
 
 export interface HomeProps {
-  claimData?: Claim[]
+  scenarioContent: ScenarioContent
   loading: boolean
 }
 
-export default function Home({ claimData, loading }: HomeProps): ReactElement {
+export default function Home({ scenarioContent, loading }: HomeProps): ReactElement {
   const { t } = useTranslation('common')
+
+  const statusContent: ClaimStatusContent = {
+    statusDescription: 'claim-status.generic-all-clear',
+    nextSteps: [
+      'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
+      'Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.',
+    ],
+  }
+
+  const detailsContent: ClaimDetailsContent = {
+    programType: 'Unemployment Insurance (UI)',
+    benefitYear: '3/21/2020 - 3/20/2021',
+    claimBalance: '$0.00',
+    weeklyBenefitAmount: '$111.00',
+    lastPaymentIssued: '4/29/2021',
+    extensionType: 'Tier 2 Extension',
+    extensionEndDate: '5/22/2021',
+  }
 
   return (
     <Container fluid className="index">
@@ -32,9 +50,9 @@ export default function Home({ claimData, loading }: HomeProps): ReactElement {
       </Head>
       <WorkInProgress />
       <Header />
-      <Main loading={loading} />
+      <Main loading={loading} statusContent={statusContent} detailsContent={detailsContent} />
       <Footer />
-      {console.dir({ claimData })} {/* @TODO: Remove. For development purposes only. */}
+      {console.dir({ scenarioContent })} {/* @TODO: Remove. For development purposes only. */}
     </Container>
   )
 }
@@ -45,7 +63,7 @@ export const getServerSideProps: GetServerSideProps = async ({ req, locale }) =>
   logger.info(req)
 
   // Make the API request and return the data.
-  const claimData: Claim = await queryApiGateway(req)
+  const claimData = await queryApiGateway(req)
 
   // Run business logic to determine the current scenario.
   const scenarioContent = getScenarioContent(claimData)
@@ -53,7 +71,7 @@ export const getServerSideProps: GetServerSideProps = async ({ req, locale }) =>
   // Return Props.
   return {
     props: {
-      claimData: [claimData],
+      scenarioContent: scenarioContent,
       loading: false,
       ...(await serverSideTranslations(locale || 'en', ['common', 'header', 'footer'])),
     },
